+++ src/interface/ifcommon.h
@@ -25,7 +25,7 @@
 //
 //
 interface IUnk {
-  virtual long IFCALL QueryInterface(REFIID, void**) = 0;
+  virtual int32_t IFCALL QueryInterface(REFIID, void**) = 0;
   virtual uint32_t IFCALL AddRef() = 0;
   virtual uint32_t IFCALL Release() = 0;
 };
